refactor: centralize Base declarative
