+++ backend/models.py
@@ -13,7 +13,7 @@ from sqlalchemy import (
 )
 from sqlalchemy.dialects.postgresql import UUID, JSONB
 from sqlalchemy.sql import func
-from .database import Base
+from backend.db_base import Base
 
 
 class Kingdom(Base):
