+++ tests/test_forgot_password_router.py
@@ -4,7 +4,7 @@ import uuid
 from sqlalchemy import create_engine, text
 from sqlalchemy.orm import sessionmaker
 
-from backend.database import Base
+from backend.db_base import Base
 from backend.models import User, Notification
 from backend.routers import forgot_password as fp
