+++ tests/test_leaderboard_router.py
@@ -2,7 +2,7 @@ import asyncio
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 
-from backend.database import Base
+from backend.db_base import Base
 from backend.models import Alliance, AllianceWar, AllianceWarScore
 from backend.routers import leaderboard
