+++ tests/test_notifications_router.py
@@ -3,7 +3,7 @@ from datetime import datetime, timedelta
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 
-from backend.database import Base
+from backend.db_base import Base
 from backend.models import Notification, User
 from backend.routers.notifications import (
     NotificationAction,
